requirements.txt
fastapi
requests
uvicorn
opentelemetry-api
opentelemetry-sdk
opentelemetry-exporter-otlp
opentelemetry-instrumentation-fastapi
httpx
numpy
ping3
